[Update] Update folder names and namespace names to match the adapter's name
