+++ dbt/adapters/fabricserverless/connection_manager.py
@@ -1,4 +1,4 @@
-from dbt.adapters.fabric import FabricConnectionManager
+from dbt.adapters.fabric.fabric_connection_manager import FabricConnectionManager
 
 class FabricServerlessConnectionManager(FabricConnectionManager):
     TYPE = "fabricserverless"
