+++ dbt/include/fabricserverless/macros/adapters/apply_grants.sql
