+++ dbt/include/fabricserverless/macros/adapters/columns.sql
