+++ dbt/include/fabricserverless/macros/adapters/indexes.sql
