+++ dbt/include/fabricserverless/macros/adapters/metadata.sql
