+++ dbt/include/fabricserverless/macros/adapters/persist_docs.sql
