+++ dbt/include/fabricserverless/macros/adapters/relation.sql
