+++ dbt/include/fabricserverless/macros/adapters/schema.sql
