+++ dbt/include/fabricserverless/macros/materializations/models/table/columns_spec_ddl.sql
