+++ dbt/include/fabricserverless/macros/materializations/models/table/create_table_as.sql
