+++ dbt/include/fabricserverless/macros/materializations/models/view/create_view_as.sql
