+++ dbt/include/fabricserverless/macros/materializations/snapshots/snapshot.sql
