+++ dbt/include/fabricserverless/macros/utils/any_value.sql
