+++ dbt/include/fabricserverless/macros/utils/array_construct.sql
