+++ dbt/include/fabricserverless/macros/utils/cast_bool_to_text.sql
