+++ dbt/include/fabricserverless/macros/utils/concat.sql
