+++ dbt/include/fabricserverless/macros/utils/date_trunc.sql
