+++ dbt/include/fabricserverless/macros/utils/hash.sql
