+++ dbt/include/fabricserverless/macros/utils/last_day.sql
