+++ dbt/include/fabricserverless/macros/utils/length.sql
