+++ dbt/include/fabricserverless/macros/utils/listagg.sql
