+++ dbt/include/fabricserverless/macros/utils/position.sql
