+++ dbt/include/fabricserverless/macros/utils/safe_cast.sql
