+++ dbt/include/fabricserverless/macros/utils/timestamps.sql
